Move placement_validations
